% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toJson.R
\name{as.diaJson.diaClassificationList}
\alias{as.diaJson.diaClassificationList}
\title{Turn object to diatech JSON}
\usage{
\method{as.diaJson}{diaClassificationList}(cls, mes, nb, ...)
}
\arguments{
\item{nb}{list with the neighbouring polygons. Gotten from the function generateMaps in this package.}

\item{...}{other paramters passing to methods.}

\item{x}{object to be turned into JSON.}

\item{labels}{data.frame containg the id column and extra variables to be added to the JSON.}
}
\description{
Turn object to diatech JSON. See specific method
}
\examples{

\dontrun{

library(diaMeasures)

data(bourciezDiaData)

set <- bourciezDiaData$lonlat

measure <- getMeasures(set, 'euclidean')
classification <- getClassification(measure, 'classic', 'Med', '107', 2:4)
result <- as.diaJson(classification, measure)

## statistics

measure <- getMeasures(set, 'euclidean')
stat <- getStatistic(measure, stat = 'sd')
classification <- getClassification(stat, 'classic', 'Med', 2:4)
json <- as.diaJson(classification, stat)

## with boundary

classification <- getClassification(measure, 'classic', 'Med', '107', 2:4, nb = bourciezDiaData$nb)
json <- as.diaJson(classification, bourciezDiaData$nb)

}

}

