% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toJson.R
\name{as.diaJson}
\alias{as.diaJson}
\title{Turn object to diatech JSON}
\usage{
as.diaJson(x, ...)
}
\arguments{
\item{x}{object to be turned into JSON}

\item{...}{other paramters passing to methods}
}
\description{
Turn object to diatech JSON. See specific method
}

