% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\docType{data}
\name{bourciezDiaData}
\alias{bourciezDiaData}
\title{Bouricez dataset}
\description{
Bourciez dataset to be used in the examples.
List containing an object of class diaLongLat and a object containing the neighbours.
}

