\name{diatechServer-package}
\alias{diatechServer-package}
\alias{diatechServer}
\docType{package}
\title{
\packageTitle{diatechServer}
}
\description{
\packageDescription{diatechServer}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{diatechServer}
\packageIndices{diatechServer}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{diatechServer}

Maintainer: \packageMaintainer{diatechServer}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\examples{
## no examples
}
