% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generateMap.R
\name{generateMaps}
\alias{generateMaps}
\title{Generate spatial content for Diatech}
\usage{
generateMaps(set, bounds)
}
\arguments{
\item{set}{list containing the data to create the polygons}

\item{bounded}{is there a boundary to clip the polygons}
}
\description{
This function will generate the maps and spatial content used in Diatech
}
\examples{
data(polydata)

## bourciez bounded
set <- polydata$bourciez$bourciez
bounds <- polydata$bourciez$bourciezBounds
bourciezMap <- generateMaps(set, bounds)

## dialeb Bounded
set <- polydata$dialeb$dialeb
bounds <- polydata$dialeb$dialebBounds
dialebMap <- generateMaps(set, bounds)

## bourciez non bounded
set <- polydata$bourciez$bourciez
bourciezMap <- generateMaps(set)
}

