% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{getClassification.diaMeasure}
\alias{getClassification.diaMeasure}
\title{Compute Classifications}
\usage{
\method{getClassification}{diaMeasure}(measure, type = c("classic"), method,
  ids, ns, nb, ...)
}
\arguments{
\item{measure}{object to pass to other methods.}

\item{type}{type of classification requried.}

\item{method}{classification method to be used: Med, MinMwMax, MedMw.}

\item{ids}{ids from the measure for which the classification method must be applied.}

\item{ns}{integer vector indicating in how many groups must the classification happen.}

\item{nb}{optional neighbourhood object returned from 'generateMaps'.}

\item{...}{unused}
}
\description{
Apply classification methods related to the Diatech project.
}
\examples{

library(diaMeasures)
data(bourciezDiaData)

set <- bourciezDiaData$lonlat
labels <- set[c('id', 'location')]
measure <- getMeasures(set, 'euclidean')

result <- getClassification(measure, 'classic', 'Med', '107', 2:4)

result2 <- getClassification(measure, 'classic', 'Med', '107', 2:4, bourciezDiaData$nb)

}

