% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{getClassification}
\alias{getClassification}
\title{Compute Classifications}
\usage{
getClassification(measure, ...)
}
\arguments{
\item{measure}{object to pass to other methods}

\item{...}{other arguments passing to methods}
}
\description{
Apply classification methods related to the Diatech project.
}

