% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/measures.R
\name{getMeasures.diaLongLat}
\alias{getMeasures.diaLongLat}
\title{Get geographical distance measures}
\usage{
\method{getMeasures}{diaLongLat}(set, measure = "euclidean")
}
\arguments{
\item{set}{object of class diaGeoDistance}

\item{measure}{measure to be used}

\item{id}{numeric of length one indicating for which location id compute the distance. If missing all will be computed.}
}
\description{
Get geographical distance measures
}
\examples{
data(bourciezDiaData)
set <- bourciezDiaData$lonlat
measure <- 'euclidean'

allEuclideanDistances <- getMeasures(set, 'euclidean')
}

