% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{getMeasures}
\alias{getMeasures}
\title{Obtain get the measures}
\usage{
getMeasures(set, measure, ...)
}
\arguments{
\item{set}{object for method to be applied}

\item{measure}{measure to be used}

\item{...}{arguments passing to other methods}
}
\description{
Applies certain ad-hoc measures used in Dialektometrics.
}

