% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{getStatistic.diaMeasure}
\alias{getStatistic.diaMeasure}
\title{Compute Statistics}
\usage{
\method{getStatistic}{diaMeasure}(measure, stat = c("sd", "min", "max",
  "median", "skewness", "cor"), measure2, ...)
}
\arguments{
\item{measure}{object of class diaMeasures.}

\item{stat}{statistic to be calculated.}

\item{measure2}{another object of class diaMeasures used for the correlation statistic.}

\item{...}{other arguments passing to methods (currently unused)}
}
\description{
Get statistics for diaMeasures.
}
\examples{

data(bourciezDiaData)
set <- bourciezDiaData$lonlat
measure <- getMeasures(set, 'euclidean')

sd <- getStatistic(measure, 'sd')
minm <- getStatistic(measure, 'min')
maxm <- getStatistic(measure, 'max')
medianm <- getStatistic(measure, 'median')
skewness <- getStatistic(measure, 'skewness')
correlation <- getStatistic(measure, 'cor', measure)

}

