% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{getStatistic}
\alias{getStatistic}
\title{Compute Statitics.}
\usage{
getStatistic(measure, stat, ...)
}
\arguments{
\item{measure}{object to pass to methods}

\item{stat}{statistic to be calculated}

\item{...}{other arguments passing to methods}
}
\description{
Get statistics related to the Diatech project
}

