% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toJson.R
\name{histBreaks}
\alias{histBreaks}
\title{Histogram breaks and density}
\usage{
histBreaks(x)
}
\description{
Returns the histogram breaks and  density estimation base on a diaClassification for the
graphical representation.
}

