% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{loadDiatech}
\alias{loadDiatech}
\title{Prototype of load data}
\usage{
loadDiatech(database, user, password, parts, type = c("geodist", "lingdata",
  "lemmadata"))
}
\arguments{
\item{database}{database name.}

\item{user}{database user.}

\item{password}{password for database.}

\item{type}{data type to be queried.}

\item{anstype}{'1' for orthographic answer, '2' for phonetic answer, '3' for all answer.}

\item{proid}{project id to be queried}

\item{linDomain}{linguistic domain to be used.}
}
\description{
This function loads data from the database.
}
\examples{

## sample call from server
\dontrun{
  set <- load('thedatabase', 'user', 'thepassword', '1', type = 'geodist')
}

}

