% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generateMap.R
\name{readSpatialDataset}
\alias{readSpatialDataset}
\title{Read Spatial Data}
\usage{
readSpatialDataset(con, proId)
}
\arguments{
\item{con}{Connection to the Diatech database.}

\item{proId}{Project Id from which to be taken the data.}
}
\description{
Reads spatial data from the Diatech database
}
\examples{
\dontrun{
## Not run:

con <- dbConnect(dbDriver("MySQL"), user = 'user', password = 'password', 'diatech')
tb <- readSpatialDataset(con, '179')

## End(**Not run**)
}
}

