% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generateMap.R
\name{writeMaps}
\alias{writeMaps}
\title{Write generated maps}
\usage{
writeMaps(map, jpegPath, polySvgPath)
}
\description{
Write generated maps
}
\examples{
\dontrun{ ## Not run:

data(polydata)

## bourciez bounded
set <- polydata$bourciez$bourciez
bounds <- polydata$bourciez$bourciezBounds
map <- generateMaps(set, bounds)
writeSVGmaps(map, 'map.jpeg')


## End(**Not run**)
}
}

