% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMap.R
\name{generateMaps}
\alias{generateMaps}
\title{Generate spatial content for Diatech}
\usage{
generateMaps(set, bounds)
}
\arguments{
\item{set}{list containing the data to create the polygons, see examples.}

\item{bounds}{Boundary matrix, see examples.}
}
\description{
This function will generate the maps and spatial content used in Diatech
}
\examples{

data(polydata)

## bourciez bounded
set <- polydata$bourciez$bourciez
names(set) <- c('id', 'x', 'y')
bounds <- polydata$bourciez$bourciezBounds
bourciezMap <- generateMaps(set, bounds)

## dialeb Bounded
set <- polydata$dialeb$dialeb
names(set) <- c('id', 'x', 'y')
bounds <- polydata$dialeb$dialebBounds
bounds <- polydata$dialeb$dialebBounds[c(3, 2, 1, 4, 5, 6, 7, 8)]
dialebMap <- generateMaps(set, bounds)

## bourciez non bounded
set <- polydata$bourciez$bourciez
names(set) <- c('id', 'x', 'y')
bourciezMap <- generateMaps(set)

}

