% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{getMeasures.diaLongLat}
\alias{getMeasures.diaLongLat}
\title{Get geographical distance measures}
\usage{
\method{getMeasures}{diaLongLat}(set, labelsCol = "id", ...)
}
\arguments{
\item{set}{object of class diaGeoDistance}

\item{labelsCol}{column name of set that will be used as labels}

\item{...}{arguments passing to other methods}
}
\description{
Get geographical distance measures
}
\examples{

data(bourciezDiaData)
set <- bourciezDiaData$lonlat
measure <- 'euclidean'

allEuclideanDistances <- getMeasures(set, 'euclidean')

}

