% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/measures.R
\name{getMeasures}
\alias{getMeasures}
\title{Obtain get the measures}
\usage{
getMeasures(set, measure, ...)
}
\description{
Applies certain ad-hoc measures used in Dialektometrics.
}
\details{
The available classification methods are Med MinMwMax and MedMw.
}

