% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.R
\name{load}
\alias{load}
\title{Prototype of load data}
\usage{
load(database, user, password, anstype, proid, linDomain, type = c("geodist",
  "lingdata", "lemmadata"))
}
\description{
This function loads data from the database. Its currently ineficients and has to be rewritten
with subqueries.
}
\examples{
## sample call from server
\dontrun{
  set <- load('thedatabase', 'user', 'thepassword', '1', type = 'geodist')
}
}

