% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMap.R
\name{writeMaps}
\alias{writeMaps}
\title{Write generated maps}
\usage{
writeMaps(map, neighPath, jpegPath, polySvgPath, centSvgPath, honeSvgPath)
}
\arguments{
\item{map}{resulting from generateMaps function.}

\item{neighPath}{path where the binary data for neighbours will be saved.}

\item{jpegPath}{path where the jpeg map will be saved.}

\item{polySvgPath}{path where the svg with the Polygon map will be saved.}

\item{centSvgPath}{path where the svg with the Centroid map will be saved.}

\item{honeSvgPath}{path where the svg witht he Honecomb map will be saved.}
}
\description{
Write generated maps
}
\examples{

\dontrun{ ## Not run:

data(polydata)

## bourciez bounded
set <- polydata$bourciez$bourciez
bounds <- polydata$bourciez$bourciezBounds
names(set) <- c('id', 'x', 'y')
map <- generateMaps(set, bounds)
writeMaps(map, 'nb.rda', 'map.jpeg', 'polygon.svg', 'cetroid.svg', 'honeycomb.svg')

## dialeb Bounded
set <- polydata$dialeb$dialeb
names(set) <- c('id', 'x', 'y')
bounds <- polydata$dialeb$dialebBounds
map <- generateMaps(set, bounds)
writeMaps(map, 'nb.rda', 'map.jpeg', 'polygon.svg', 'cetroid.svg', 'honeycomb.svg')

## End(**Not run**)
}

}

